Conflicts:
	proyecto-rem-online/sencha-app/classic/src/ux/util/Constants.js
	proyecto-rem-online/sencha-app/resources/locale/messages_es.properties
	recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/api/ActivoApi.java
	sql/pitertul

Former-commit-id: 1ed4437f4ed3507990a79439b66a76887eb7d972

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/activo/ActivoManager.java
@@ -39,11 +39,13 @@ import es.pfsgroup.framework.paradise.utils.BeanUtilNotNull;
 import es.pfsgroup.framework.paradise.utils.DtoPage;
 import es.pfsgroup.plugin.recovery.coreextension.utils.api.UtilDiccionarioApi;
 import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.DDUnidadPoblacional;
+import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.NMBValoracionesBien;
 import es.pfsgroup.plugin.rem.activo.dao.ActivoDao;
 import es.pfsgroup.plugin.rem.adapter.ActivoAdapter;
 import es.pfsgroup.plugin.rem.adapter.GenericAdapter;
 import es.pfsgroup.plugin.rem.api.ActivoApi;
 import es.pfsgroup.plugin.rem.api.TrabajoApi;
+import es.pfsgroup.plugin.rem.api.UvemManagerApi;
 import es.pfsgroup.plugin.rem.factory.TabActivoFactoryApi;
 import es.pfsgroup.plugin.rem.model.Activo;
 import es.pfsgroup.plugin.rem.model.ActivoAdjuntoActivo;
@@ -60,6 +62,7 @@ import es.pfsgroup.plugin.rem.model.ActivoOferta;
 import es.pfsgroup.plugin.rem.model.ActivoPropietarioActivo;
 import es.pfsgroup.plugin.rem.model.ActivoProveedor;
 import es.pfsgroup.plugin.rem.model.ActivoSituacionPosesoria;
+import es.pfsgroup.plugin.rem.model.ActivoTasacion;
 import es.pfsgroup.plugin.rem.model.ActivoValoraciones;
 import es.pfsgroup.plugin.rem.model.DtoActivoDatosRegistrales;
 import es.pfsgroup.plugin.rem.model.DtoActivoFichaCabecera;
@@ -79,6 +82,7 @@ import es.pfsgroup.plugin.rem.model.DtoOfertaActivo;
 import es.pfsgroup.plugin.rem.model.DtoPrecioVigente;
 import es.pfsgroup.plugin.rem.model.DtoPropuestaActivosVinculados;
 import es.pfsgroup.plugin.rem.model.DtoPropuestaFilter;
+import es.pfsgroup.plugin.rem.model.DtoTasacion;
 import es.pfsgroup.plugin.rem.model.ExpedienteComercial;
 import es.pfsgroup.plugin.rem.model.GastosExpediente;
 import es.pfsgroup.plugin.rem.model.Oferta;
@@ -145,6 +149,9 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 
 	@Autowired
 	private VisitaDao visitasDao;
+	
+	@Autowired
+	private UvemManagerApi uvemManagerApi;
 
 	@Override
 	public String managerName() {
@@ -439,12 +446,29 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 	@Override
 	@Transactional(readOnly = false)
 	public boolean deleteValoracionPrecio(Long id) {
-
+		
+		return deleteValoracionPrecioConGuardadoEnHistorico(id,true);
+	}
+	
+	@Override
+	@Transactional(readOnly = false)
+	public boolean deleteValoracionPrecioConGuardadoEnHistorico(Long id, Boolean guardadoEnHistorico) {
+		
 		Filter filtro = genericDao.createFilter(FilterType.EQUALS, "id", id);
 		ActivoValoraciones activoValoracion = genericDao.get(ActivoValoraciones.class, filtro);
-		saveActivoValoracionHistorico(activoValoracion);
-		// genericDao.deleteById(ActivoValoraciones.class, id);
-		activoDao.deleteValoracionById(id);
+		
+		if(guardadoEnHistorico) {
+			saveActivoValoracionHistorico(activoValoracion);
+			activoDao.deleteValoracionById(id);
+		}
+		else if(!Checks.esNulo(activoValoracion.getGestor()) && !adapter.getUsuarioLogado().equals(activoValoracion.getGestor())) {
+			//Si el usuario logado es distinto al que ha creado la valoracion, no puede borrarla sin historico
+			return false;
+		}
+		else {
+			//Al anular el precio vigente, se hace un borrado lógico, y no se inserta en el histórico
+			genericDao.deleteById(ActivoValoraciones.class, id);
+		}
 
 		return true;
 	}
@@ -1140,6 +1164,12 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 				Filter proveedorFiltro = genericDao.createFilter(FilterType.EQUALS, "id", Long.parseLong(dto.getMediador()));
 				ActivoProveedor proveedor = genericDao.get(ActivoProveedor.class, proveedorFiltro);
 				beanUtilNotNull.copyProperty(historicoMediador, "mediadorInforme", proveedor);
+				
+				// Asignar el nuevo proveedor de tipo mediador al activo, informacion comercial.
+				if(!Checks.esNulo(activo.getInfoComercial())) {
+					beanUtilNotNull.copyProperty(activo.getInfoComercial(), "mediadorInforme", proveedor);
+					genericDao.save(Activo.class, activo);
+				}
 			}
 
 			genericDao.save(ActivoInformeComercialHistoricoMediador.class, historicoMediador);
@@ -1636,5 +1666,94 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		}
 		
 		return contador;
+	}
+
+
+	@Override
+	@Transactional(readOnly = false)
+	public Boolean solicitarTasacion(Long idActivo) {
+		
+		try{
+			// Se especifica bankia por que tan solo se va a poder demandar la tasación desde bankia.
+			int tasacionID = uvemManagerApi.ejecutarSolicitarTasacion(idActivo, adapter.getUsuarioLogado().getNombre(), "BANKIA");
+			if(!Checks.esNulo(tasacionID)){
+				Activo activo = activoDao.get(idActivo);
+				
+				if(!Checks.esNulo(activo)) {
+					// Generar un 'BIE_VALORACION' con el 'BIEN_ID' del activo.
+					NMBValoracionesBien valoracionBien = new NMBValoracionesBien();
+					beanUtilNotNull.copyProperty(valoracionBien, "bien", activo.getBien());
+					valoracionBien = genericDao.save(NMBValoracionesBien.class, valoracionBien);
+					
+					if(!Checks.esNulo(valoracionBien)) {
+						// Generar una tasacion con el ID de activo y el ID de la valoracion del bien.
+						ActivoTasacion tasacion = new ActivoTasacion();
+						
+						beanUtilNotNull.copyProperty(tasacion, "idExterno", tasacionID);
+						beanUtilNotNull.copyProperty(tasacion, "activo", activo);
+						beanUtilNotNull.copyProperty(tasacion, "valoracionBien", valoracionBien);
+						
+						genericDao.save(ActivoTasacion.class, tasacion);
+					}
+				}
+			}
+		}catch(Exception e){
+			e.printStackTrace();
+			return false;
+		}
+		
+		return true;
+	}
+
+	@Override
+	public DtoTasacion getSolicitudTasacionBankia(Long id) {
+		ActivoTasacion activoTasacion = activoDao.getActivoTasacion(id);
+		DtoTasacion dtoTasacion = new DtoTasacion();
+		if(!Checks.esNulo(activoTasacion)) {
+			
+			try {
+				beanUtilNotNull.copyProperty(dtoTasacion, "id", activoTasacion.getId());
+				beanUtilNotNull.copyProperty(dtoTasacion, "fechaSolicitudTasacion", activoTasacion.getAuditoria().getFechaCrear());
+				beanUtilNotNull.copyProperty(dtoTasacion, "gestorSolicitud", activoTasacion.getAuditoria().getUsuarioCrear());
+				beanUtilNotNull.copyProperty(dtoTasacion, "externoID", activoTasacion.getIdExterno());
+			} catch (IllegalAccessException e) {
+				e.printStackTrace();
+			} catch (InvocationTargetException e) {
+				e.printStackTrace();
+			}
+		}
+		
+		return dtoTasacion;
+	}
+	
+	@Override
+	@BusinessOperationDefinition("activoManager.comprobarActivoComercializable")
+	public Boolean comprobarActivoComercializable(Long idActivo) {
+		PerimetroActivo perimetro = this.getPerimetroByIdActivo(idActivo);
+		
+		return perimetro.getAplicaComercializar() == 1 ? true : false;
+	}
+   
+	@Override
+	@BusinessOperationDefinition("activoManager.comprobarObligatoriosDesignarMediador")
+	public String comprobarObligatoriosDesignarMediador(Long idActivo) throws Exception {
+
+		Activo activo = this.get(idActivo);
+		String mensaje = new String();
+
+		// Validaciones datos obligatorios correspondientes a Publicacion / Informe comercial
+		// del activo
+		// Validación mediador
+		if (Checks.esNulo(activo.getInfoComercial()) || Checks.esNulo(activo.getInfoComercial().getMediadorInforme())) {
+			mensaje = mensaje.concat(
+					messageServices.getMessage("tramite.admision.DesignarMediador.validacionPre.mediador"));
+		}
+
+		if (!Checks.esNulo(mensaje)) {
+			mensaje = messageServices.getMessage("tramite.admision.DesignarMediador.validacionPre.debeInformar")
+					.concat(mensaje);
+		}
+
+		return mensaje;
 	}
 }
